pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hopper"
version = "0.1.0"
description = "Unified Python library for AI model API calls"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.0.0"]
google = ["google-genai>=0.8.0"]
together = ["together>=1.3.0"]
perplexity = ["openai>=1.0.0"]
grok = ["openai>=1.0.0"]
all = [
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "google-genai>=0.8.0",
    "together>=1.3.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.hatch.build.targets.wheel]
include = ["hopper/**"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
